refactoring the code for the paged (sorry boys)
